requirements.txt
Django==2.2
django-extensions==2.2.9
django-filter==2.2.0
djangorestframework==3.11.0
drf-yasg==1.17.1
markdown==3.2.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
swagger-spec-validator==2.5.0
django-environ
